Update README.md
Wallet Verification Contract
This is a smart contract for verifying the ownership of a wallet address and issuing a token to the verified address.

Getting Started
To use this contract, you'll need to have a working Ethereum development environment set up. This includes:

An Ethereum client (such as Geth or Parity)
An Ethereum development framework (such as Truffle)
A text editor or IDE for editing Solidity code
Installing
Clone the repository to your local machine:

bash
Copy code
git clone https://github.com/<username>/<repository>.git
Usage
To use the wallet verification contract, you'll need to deploy it to an Ethereum network. You can do this using a tool such as Truffle or Remix.

Once the contract is deployed, you can interact with it by calling the verify function, passing in the address you want to verify as an argument:

php
Copy code
function verify(address recipient) public returns (bool)
This function will verify the specified address and issue a token to it if it hasn't already been verified.

Built With
Solidity - The programming language used
Truffle - The development framework used
Remix - The online Solidity IDE used for testing and deployment
Authors
John Doe - Initial work
License
This project is licensed under the MIT License - see the LICENSE file for details.